+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add util `\CfdiUtils\Utils\CurrencyDecimals`, help to work with decimals by currency
 - Improve `CfdiUtils\Validate\Cfdi33\Standard\ComprobanteDecimalesMoneda` with previous class
 - Add util `\CfdiUtils\Utils\Rfc`, help to work with strict RFC validations
+    - Add validator `\CfdiUtils\Validate\Cfdi33\Standard\ReceptorRfc`
 - Add docblocks to `CfdiUtils\Cfdi`
 - Building:
     - Add .phplint.yml to export-ignore (standard line)